requirements.txt
# GitHub Actions용 필수 패키지
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
selenium
webdriver-manager
pyperclip
holidays
pytz
matplotlib
pandas
schedule 